requirements.txt
flask==2.3.3
requests==2.31.0
psycopg2-binary==2.9.3
lxml==4.9.3
flask_cors